api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["malo.girard <malo.girard@jellysmack.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
fastapi = "^0.105.0"
opensearch-py = "^2.4.2"
uvicorn = "^0.25.0"
qdrant-client = "^1.7.0"
fastembed = "^0.1.3"
boto3 = "^1.34.5"
pyscopg2 = "^66.0.2"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
